Use optional slug parameter for version
