+++ sites/svelte.dev/src/lib/docs/types.d.ts
@@ -1,0 +1,6 @@
+export interface DocsVersion {
+	git_ref: string;
+	version: string;
+	title: string;
+	description: string;
+}
