+++ sites/svelte.dev/src/lib/server/docs/types.d.ts
@@ -24,10 +24,3 @@ export type Page = {
 	content: string;
 	sections: Section[];
 };
-
-export interface DocsVersion {
-	git_ref: string;
-	version: string;
-	title: string;
-	description: string;
-}
