+++ sites/svelte.dev/src/params/version.js
@@ -1,0 +1,6 @@
+import versions from '$lib/docs/versions.js';
+
+/**
+ * @type {import('@sveltejs/kit').ParamMatcher}
+ */
+export const match = (param) => versions.find((v) => v.version === param) != null;
