+++ sites/svelte.dev/src/routes/docs/[[version=version]]/[slug]/+page.server.js
@@ -6,11 +6,8 @@ export const prerender = true;
 /**
  * @type {import('./$types').PageServerLoad}
  */
-export const load = async ({ params, url }) => {
-	const processed_page = await get_parsed_docs(
-		await get_docs_data(url.searchParams.get('version')),
-		params.slug
-	);
+export const load = async ({ params }) => {
+	const processed_page = await get_parsed_docs(await get_docs_data(params.version), params.slug);
 
 	if (!processed_page) error(404);
