+++ sites/svelte.dev/src/routes/docs/[[version=version]]/[slug]/+page.svelte
